Restructed native folders
