+++ src/natives/array.hpp
@@ -1,0 +1,10 @@
+#ifndef _SG_CPP_NATIVES_ARRAY_HPP
+#define _SG_CPP_NATIVES_ARRAY_HPP
+
+#include "../value.hpp"
+
+namespace Natives {
+    Values::Value create_array_namespace();
+};
+
+#endif
